fix(web): share the public country default
