+++ apps/web/src/lib/case-law-route.ts
@@ -1,6 +1,41 @@
 import { panic } from "better-result";
 
+import {
+  isPublicCaseLawCountry,
+  PUBLIC_CASE_LAW_COUNTRIES,
+} from "@stll/api-contract/case-law-launch-readiness";
+import type {
+  CaseLawBrowserCountry,
+  PublicCaseLawCountry,
+} from "@stll/api-contract/case-law-launch-readiness";
+import type { UiLocale } from "@stll/locales";
 import { stripDiacriticsForSlug } from "@stll/text-normalize";
+
+const DEFAULT_COUNTRY_BY_LOCALE = {
+  ar: null,
+  cs: "CZE",
+  de: null,
+  en: null,
+  es: null,
+  et: null,
+  fr: null,
+  hu: null,
+  lt: null,
+  lv: null,
+  pl: "POL",
+  "pt-BR": null,
+  sk: "SVK",
+} as const satisfies Record<UiLocale, CaseLawBrowserCountry | null>;
+
+/** Pick a launch-ready public country from the UI locale, then list order. */
+export const defaultCaseLawCountryForLocale = (
+  locale: UiLocale,
+): PublicCaseLawCountry | null => {
+  const localeCountry = DEFAULT_COUNTRY_BY_LOCALE[locale];
+  return localeCountry !== null && isPublicCaseLawCountry(localeCountry)
+    ? localeCountry
+    : (PUBLIC_CASE_LAW_COUNTRIES.at(0) ?? null);
+};
 
 const UUID_REGEX =
   /^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$/u;
