+++ apps/web/src/routes/law/cases/index.tsx
@@ -30,7 +30,6 @@ import { Button } from "@stll/ui/button";
 import { Skeleton } from "@stll/ui/skeleton";
 
 import {
-  defaultCaseLawCountryForLocale,
   PUBLIC_CASE_LAW_COUNTRIES,
   publicCaseLawCountryFromParam,
   toCaseLawCountryParam,
@@ -63,6 +62,7 @@ import { getMessageLocale } from "@/i18n/i18n-store";
 import {
   createCaseLawDecisionPath,
   createCaseLawDecisionRouteParams,
+  defaultCaseLawCountryForLocale,
 } from "@/lib/case-law-route";
 import { detached } from "@/lib/detached";
 import { pageTitle } from "@/lib/page-title";
