+++ docs/module-ownership.md
@@ -39,7 +39,7 @@ reason, in the same table.
 | `bounded-concurrency` — Running an async operation over a list with a bounded number in flight | `packages/concurrency/` | none | A windowed `Promise.all` over slices is the shape everyone reaches for and it is not a concurrency bound: the window refills only once its slowest member settles, so effective concurrency decays to each slice's tail and drops to zero for whatever the caller does between slices. `mapWithConcurrency` returns every result and keeps the pool full throughout. `streamWithConcurrency` yields each result in input order as it is ready, and its `lookAhead` decides whether the pool refills on completion or on consumption: at the default of zero a settled result holds its slot, so the pool slides and residency stays at `limit`; a caller that wants work to continue through a slow item and through its own per-result work has to ask for look-ahead and pay for up to `limit + lookAhead` resident results. The stream observes each settlement at the pool, so a rejection behind a slower item cannot surface as an unhandled rejection, and nothing it started outlives its consumer: closing the stream starts nothing more and waits for what is already running. |
 | `case-law-adapter-manifest` — Case-law source declarations | `apps/api/src/lib/legal-search/adapter-manifest.ts`, `apps/api/src/lib/case-law/ecli-court-codes.ts` | none | One total map binds every adapter key to its source name, jurisdiction, known ECLI court codes, declared text sentinels, and date range. Each jurisdiction selection also carries its declared docket grammar. Adapters read the declaration, and the runner reads the resulting total registry, so adding a source requires one complete entry. |
 | `case-law-docket-grammar` — Parsing and comparing decision docket identifiers | `packages/api-contract/src/decision-docket-grammar.ts`, `packages/api-contract/src/decision-query-intent.ts` | none | One total jurisdiction map recognizes docket syntax and returns normalized display and comparison forms. Search intent classification and exact result matching use that same parser, so a spelling cannot be classified under one rule and compared under another. |
-| `case-law-launch-readiness` — Selecting countries exposed by public case-law surfaces | `packages/api-contract/src/case-law-launch-readiness.ts`, `packages/api-contract/src/launch-readiness.json` | none | One checked-in inclusion list carries complete readiness evidence for each public country. The shared parser rejects incomplete or ambiguous rows, and web and API consumers use the resulting country boundary. |
+| `case-law-launch-readiness` — Selecting countries exposed by public case-law surfaces | `packages/api-contract/src/case-law-launch-readiness.ts`, `packages/api-contract/src/launch-readiness.json`, `apps/web/src/lib/case-law-route.ts` | none | One checked-in inclusion list carries complete readiness evidence for each public country. The shared parser rejects incomplete or ambiguous rows, and web and API consumers use the resulting country boundary. |
 | `case-law-decision-text` — Representing publisher-authored decision text at ingestion and public read boundaries | `packages/api-contract/src/case-law-text-field.ts`, `apps/api/src/lib/case-law/decision-headnote.ts`, `apps/api/src/lib/case-law/decision-headnote-schema.ts`, `apps/api/src/lib/case-law/decision-text.ts`, `apps/api/src/lib/case-law/decision-text-sql.ts` | none | The shared discriminated values make text presence and row-preview truncation explicit to API consumers. The API owner classifies source values, converts them to storage metadata, and reconstructs bounded public values; the adapter lint keeps protected keys behind that boundary. |
 | `search-total` — Declaring whether a search result total was counted | `packages/api-contract/src/search.ts`, `apps/api/src/lib/search/total-schema.ts` | none | One discriminated contract distinguishes an exact count, an estimate, and a search that did not compute a count. The API schema enforces the same closed branches at response boundaries, so consumers never infer count semantics from null or from the search implementation. |
 | `agent-input-normalization` — Reading a value a model wrote on the MCP/CLI wire: dates, date formats, numbers, booleans, locales, closed vocabularies | `packages/agent-input/src/` | global `Intl.getCanonicalLocales` | Every agent-facing surface is lenient in the same way or it is lenient in several different ways, which is worse than being strict: `4 000`, `1. 10. 2026`, `ano` and `cs_CZ` have to mean the same thing in a marker, in a tool input and in a fill value. One reader per kind auto-normalizes the spellings that carry a single meaning and returns the one ask-for-a-fix shape (`received`, `expected`, `hint`) when a spelling carries two — `01/02/2026` and a bare `1,234` are asked about, never guessed, because guessing them wrong is a wrong date or a factor of a thousand on an instrument. The rule below confines locale plausibility, whose canonical spelling is also what keeps `new Intl.DateTimeFormat` from throwing at fill time; the other kinds are held by the census in `agent-input-owner.test.ts`, which lists the pre-existing readers of our own bytes that are not agent input. |
